# target/classes/swagger/api.yml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Upgrade - Island
  description: Upgrade Island Api.

basePath: /v1

tags:
  - name: reservations
    description: Reservations CRUD
  - name: availabilities
    description: Reservation date availabilities


schemes:
  - http
  - https


produces:
  - application/json

definitions:
  reservationFieldsDto:
    $ref: ./models/ReservationFieldsDto.yml

  reservationDto:
    $ref: ./models/ReservationDto.yml

  errorResponse:
    type: object
    properties:
      message:
        type: string
      code:
        type: integer


parameters:
  reservationFieldsDto:
    name: ReservationFieldsDto
    in: body
    required: true
    schema:
      $ref: '#/definitions/reservationFieldsDto'


responses:
  Created:
    description: Resource Created, headers will contain a redirect.
  Confirmed:
    description: Operation confirmed, no content nor redirect.

  BadRequest:
    description: Bad request.
    schema:
      $ref: '#/definitions/errorResponse'

  NotFound:
    description: Resource not found.
    schema:
      $ref: '#/definitions/errorResponse'

  ValidationError:
    description: Input Validation, error will contain the details
    schema:
      $ref: '#/definitions/errorResponse'


paths:

  /availabilities:
    get:
      tags:
        - availabilities
      parameters:
        - name: startDate
          type: string
          format: date
          in: query
        - name: endDate
          type: string
          format: date
          in: query

      summary: get availabilities
      description: get availabilities
      responses:
        '200':
          description: list of availability
          schema:
            type: array
            items:
              title: availability
              type: object
              properties:
                date:
                  type: string
                  format: date
                isAvailable:
                  type: boolean

        '422':
          $ref: "#/responses/ValidationError"


  /reservations:
    post:
      operationId: createReservation
      tags:
        - reservations
      parameters:
        - $ref: '#/parameters/reservationFieldsDto'

      summary: create a reservation
      description: create a reservation
      responses:
        '200':
          schema:
            $ref: '#/definitions/reservationDto'
        '422':
          $ref: "#/responses/ValidationError"
#        '201':
#          description: "Created, with Location header pointing to the newly created resource and response body containing the resource"

  /reservations/{reservationId}:
    get:
      operationId: getReservation
      tags:
        - reservations
      parameters:
        - name: reservationId
          description: reservation unique identifier.
          type: integer
          in: path
          required: true

      summary: get a reservation
      description: get a reservation
      responses:
        '200':
          schema:
            $ref: '#/definitions/reservationDto'

    put:
      operationId: updateReservation
      tags:
        - reservations
      parameters:
        - name: reservationId
          description: reservation unique identifier.
          type: integer
          in: path
          required: true
        - $ref: '#/parameters/reservationFieldsDto'

      summary: update a reservation
      description: update a reservation
      responses:
        '200':
          schema:
            $ref: '#/definitions/reservationDto'
        '422':
          $ref: "#/responses/ValidationError"

#        '204':
#          description: "Created, with Location header pointing to the newly created resource and response body containing the resource"


    delete:
      operationId: deleteReservation
      tags:
        - reservations
      parameters:
        - name: reservationId
          description: reservation unique identifier.
          type: integer
          in: path
          required: true

      summary: delete a reservation
      description: delete a reservation
      responses:
        '204':
          description: "Reservation deleted with success"